backend/requirements.txt
fastapi
uvicorn
groq
pinecone
sentence-transformers
docling
python-multipart
pydantic
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
